Update service test to not be different in sync async mode
Before this CL the test fails because we land on the "await" when
stepping out of "helper" as "helper" is executed synchronously (i.e.
the call to await (which is translated into a call to an await-helper)
hasn't executed yet so landing on it is "correct").
This CL changes the test so "helper" is not executed synchronously by
adding an "await null;" as the first thing in the test. Once we step out
we thus have already executed the "await" and doesn't land on it.

An alternative fix would have been to pass --no-sync-async.

Change-Id: Iab6e5387960c2ebba10602f5e90630c30172c66e
Reviewed-on: https://dart-review.googlesource.com/63581
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Ryan Macnak <rmacnak@google.com>