Force tests to fail pending implementation.
Some of the expected failures passed on Windows - I think it's because they're not creating the "empty edits" on Windows (probably because something else is going wrong with finding which files need updating).

So for now, this forces them to fail (so they "pass"). I'll do some testing on Windows when working on these next week to figure out what's wrong.

Change-Id: I55cbf42f233e7f19b95733cfbd136b0ff1803607
Reviewed-on: https://dart-review.googlesource.com/63882
Reviewed-by: Devon Carew <devoncarew@google.com>
Commit-Queue: Danny Tuppeny <dantup@google.com>