Fix force-strong flag.
This has a tiny effect in reporting warnings, other than that the flags were
interpreted correctly during compilation.

The force-flag was is used to recognize when the user
explicitly said `--strong` and when it was implicit, so if we see
`--no-preview-dart-2` and we don't see `--strong`, we know we are in legacy mode
and we don't show warnigns about `--checked` and `--trust-type-annotations`.

Change-Id: I532ce4520b49e3aac1ec13898186c74585855d3e
Reviewed-on: https://dart-review.googlesource.com/63980
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>