[gn] Remove dart_test_bootfs
This doesn't look like it could have been working for a while.
It uses obsolete Fuchsia GN infrastructure that is going away.
If a test setup like this is useful it can be resurrected in a
new form that fits into the current Fuchsia ways of doing things.

Change-Id: I41e2ad5948cbf71a5e83f8e2f3235aa0dbd5c9a3
Reviewed-on: https://dart-review.googlesource.com/64022
Reviewed-by: Zach Anderson <zra@google.com>
Commit-Queue: Zach Anderson <zra@google.com>