Update dartfmt to 1.1.3.
We also need to update the prebuilt SDK to keep it in sync with this
change so the presubmit doesn't get wonky. But, in this case, it should
be fine. The change affects a rare case that I don't think occurs in
the SDK: a multi-line string inside a string interpolation, inside
another string. If that does occur, it's easy to just change that code.

Change-Id: Icdfdfb0b2e1359005d88d983e9a3cf05b0e91e48
Reviewed-on: https://dart-review.googlesource.com/65280
Reviewed-by: Jenny Messerly <jmesserly@google.com>