Postpone type/element translation until ResolutionApplier.
This CL moves type/element translation from the
_ResolutionApplierContext class to the ResolutionApplier.  This allows
types and elements to be translated in the context in which they
appear in source code, so that we can resolve type parameter
references properly even when they refer to type parameters of local
functions and function expressions.

Fixes #33722.
Fixes #33874.

Change-Id: I100f417654c53ee24e8ca4e7295761df0812a83b
Reviewed-on: https://dart-review.googlesource.com/65404
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>