Stop tracking ResolutionData.literalType in analyzer/FE integration.
In all but one case, the types that had previously been stored in
ResolutionData.literalType have become redundant; we now communicate
the necessary information in a higher fidelity way using
typeReference(), voidType(), and functionType().

There is one case that isn't redundant: variables declared in a catch
clause.  In this case the type in question really isn't a literal type
at all, it's an inferred type, so it makes more sense to store it in
ResolutionData.inferredType.

Change-Id: Ic03f6e5ba4bd91b561b3ee46b02324758763bc83
Reviewed-on: https://dart-review.googlesource.com/65482
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>