Switch to use swarm instead of sunflower for dart2js unittests.
These where the last two uses of sunflower in the repo. Sunflower sources have
been deleted from the public github repo. We can alternatively copy them here,
but for these tests it doesn't really matter what code we use.

A benefit of using swarm: this will test compiling swarm, so we detect breakages
early (and not when they hit the perf bots). A drawback of using swarm: it is
bigger and these tests will take longer.  For dump-info that might be a good
thing, though.

TBR

Change-Id: Ie976cf4250f5f7ce17147ba8b166e7a6bd429425
Reviewed-on: https://dart-review.googlesource.com/65784
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>