[vm/compiler] Bug fix on background to foreground passing.
Rationale:
Fixes a bug when JIT compiler sets is_background_optimizable
to false on a function to ensure it is no longer compiled on
the background but on the foreground. Before background
thread would directly put it back on its own queue which,
in some cases, could cause a compilation retry storm.

The new code correctly removes it from the queue so
that, with the higher usage count, the foreground thread
will soon pick it up.


https://github.com/dart-lang/sdk/issues/33664

Change-Id: I6e6041f571c2b32b13911c3b46b8688625d0834e
Reviewed-on: https://dart-review.googlesource.com/66000
Commit-Queue: Aart Bik <ajcbik@google.com>
Reviewed-by: Ryan Macnak <rmacnak@google.com>