Fix ExtractTokenPositionFromAsyncClosure on kernel
Previously (e71bd048e57) it was assumed that kernels
`script.yield_positions` was for the single function, not for the entire
script. That's not the case.
This resulted in the wrong stack-trace sometimes being produced.

This is technically tested in a service test, though that test also
tests other things and fails before getting to test this.

This CL adds a test that tests only this and fixes the problem.

Cloned from https://dart-review.googlesource.com/c/sdk/+/63741 to
untangle it from two other changes.

Bug: https://github.com/flutter/flutter/issues/17838
Change-Id: I4c24342d63865cb7cbd25c42ba686007bdbde1b1
Reviewed-on: https://dart-review.googlesource.com/66800
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Alexander Aprelev <aam@google.com>