Triage the remaining analyzer/front_end integration language_2 failures.
For most failures I've added a comment with a short description of the
failure.  In a few places I've included a reference to an issue in the
issue tracker.  I'll follow up with a CL that assigns issues to the
remaining failures (and I'll file issues if necessary).

I haven't tried to classify the "MissingCompileTimeError" or
"MissingStaticWarning" cases, since they are most likely due to
error/warning checks that haven't been implemented in the front end
yet, and that work is being tracked elsewhere.

Change-Id: I41559ed848a687767eb1018502cbbc104569a7cd
Reviewed-on: https://dart-review.googlesource.com/69880
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>