New rule: avoid_void_async. `void f() async`, suggest `Future<void>`. (dart-lang/linter#1128)
* New rule: avoid_void_async. `void f() async`, suggest `Future<void>`.

Fix dart-lang/linter#1122.

As per discussion there, this is a *new* rule, not part of void_checks,
because it is much more likely to have low false positives.

In my experince, void_checks is very strict. It serves a useful purpose
in pushing people away from the more complex Null type by making void
useful in a way Null used to be used. However, void_checks is not
required for soundness, and it seems uncommon that the theory here
misses something practical.

For that reason, make a new rule. Users interested in making void behave
as if it were Null in certain ways can enable that, while users who just
want to make sure no async functions are declared wrong can enable this.

* Test on appveyor: move Stream into dart.async main file.