Call our error events in StreamTransform.bind doc
Fixes #27739

The doc already has guidelines on things like listening to and pausing
the stream but did not mention whether or how error events should be
considered. Since a Transformer may choose to do anything with errors
metion that explicitly and encourage the behavior to be documented.

Change-Id: I77e98c0967ae0041f5e959282369aa294aa9b18a
Reviewed-on: https://dart-review.googlesource.com/71383
Commit-Queue: Nate Bosch <nbosch@google.com>
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>